refactor: Relocate enums to a dedicated package
This commit refactors the codebase by moving existing enum classes into a new `data.enums` package for better organization and code structure.

Key changes:
- Created a new package `com.jksalcedo.passvault.data.enums`.
- Moved `SortOption`, `ImportType`, and `Action` enums to the new package.
- Deleted the old `SortOption.kt` file and removed the inline `ImportType` and `Action` enums from their respective UI files.
- Updated all import statements across the application to reference the new location of these enums.
