+++ app/src/main/java/com/jksalcedo/passvault/data/enums/Action.kt
@@ -1,0 +1,5 @@
+package com.jksalcedo.passvault.data.enums
+
+enum class Action {
+    IMPORT, EXPORT
+}
