+++ app/src/main/java/com/jksalcedo/passvault/data/enums/ImportType.kt
@@ -1,0 +1,9 @@
+package com.jksalcedo.passvault.data.enums
+
+enum class ImportType {
+    PASSVAULT_JSON,
+    PASSVAULT_CSV,
+    BITWARDEN_JSON,
+    KEEPASS_CSV,
+    KEEPASS_KDBX
+}
